day14/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
tiktoken==0.5.2
requests==2.31.0
numpy==1.24.3

